Title: 최솟값 3개, Time: 959ms, Memory: 27MB, Status: Passed - Codetree

+++ 260113/최솟값 3개/top-3-smallest-number.java
@@ -2,9 +2,9 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        PriorityQueue<Integer> pq=new PriorityQueue<>();
+        PriorityQueue<Long> pq=new PriorityQueue<>();
         int n = sc.nextInt();
-        int t;
+        long t;
         for(int i =0;i<n;i++){
             t = sc.nextInt();
             pq.add(t);
@@ -12,9 +12,9 @@ public class Main {
                 System.out.println(-1);
             }
             else{
-                int a = pq.poll();
-                int b = pq.poll();
-                int c = pq.poll();
+                long a = pq.poll();
+                long b = pq.poll();
+                long c = pq.poll();
                 System.out.println(a*b*c);
                 pq.add(a);
                 pq.add(b);
